+++ CHANGELOG.md
@@ -17,6 +17,8 @@ with the `YY.MINOR.MICRO` scheme.
   Tiers can span multiple failure domains and a single cluster can have
   multiple tiers. Going forward it will be possible to specify which
   tier a table belongs to.
+- New option `picodata run -c` for specifying any configuration parameters in
+  the same naming scheme as in the config.yaml file.
 - New option `picodata run --tier` specifies whether an
   instance belongs to a tier.
 -->